+++ package.json
@@ -19,7 +19,7 @@
     "@devbookhq/splitter": "^1.3.2",
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.58",
+    "@material-ui/lab": "4.0.0-alpha.61",
     "@material-ui/pickers": "^3.3.10",
     "algoliasearch": "^4.9.1",
     "browser-image-compression": "^1.0.17",
